+++ core/server/api/canary/settings.js
@@ -2,7 +2,6 @@ const Promise = require('bluebird');
 const _ = require('lodash');
 const models = require('../../models');
 const routeSettings = require('../../services/route-settings');
-const frontendSettings = require('../../../frontend/services/settings');
 const i18n = require('../../../shared/i18n');
 const {BadRequestError} = require('@tryghost/errors');
 const settingsService = require('../../services/settings');
@@ -191,7 +190,7 @@ module.exports = {
         },
         async query(frame) {
             await routeSettings.api.setFromFilePath(frame.file.path);
-            const getRoutesHash = () => frontendSettings.getCurrentHash();
+            const getRoutesHash = () => routeSettings.api.getCurrentHash();
             await settingsService.syncRoutesHash(getRoutesHash);
         }
     },
